+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.66.0](https://github.com/AtharvaCM/vehicle-vault/compare/v1.65.13...v1.66.0) (2026-09-23)
+
+
+### Features
+
+* **tyres:** front/rear tyre positions and a two-wheel layout ([fade934](https://github.com/AtharvaCM/vehicle-vault/commit/fade934952a3c98089ebfa087f472dc5a1c4371e))
+
+## [1.65.13](https://github.com/AtharvaCM/vehicle-vault/compare/v1.65.12...v1.65.13) (2026-09-23)
+
+
+### Bug Fixes
+
+* **reminders:** count the suggested schedule from the last logged service ([fe73860](https://github.com/AtharvaCM/vehicle-vault/commit/fe738606ccdbd58fecdeceb5b4e602586f93adfb))
+
 ## [1.65.12](https://github.com/AtharvaCM/vehicle-vault/compare/v1.65.11...v1.65.12) (2026-09-23)